Factor out make_shader.rs to redulce assets.rs code
